+++ lib/gecko/helpers/association_helper.rb
@@ -12,6 +12,7 @@ module Gecko
       # @param [Symbol]   model_name
       # @param [#to_hash] options the options
       # @option options [String] :class_name Override the default class name
+      # @option options [String] :readonly (false) Whether to serialize this attribute
       #
       # @return [undefined]
       #
@@ -19,13 +20,14 @@ module Gecko
       def belongs_to(model_name, options={})
         class_name  = options[:class_name] || model_name.to_s.classify
         foreign_key = model_name.to_s.foreign_key.to_sym
+
         define_method model_name do
           id = self.attributes[foreign_key]
           if id
             @client.adapter_for(class_name).find(id)
           end
         end
-        attribute foreign_key, Integer
+        attribute foreign_key, Integer, readonly: options[:readonly]
       end
 
       # Set up a has_many relationship between two classes based on a JSON key of
@@ -39,6 +41,7 @@ module Gecko
       # @param [Symbol]   model_name
       # @param [#to_hash] options the options
       # @option options [String] :class_name Override the default class name
+      # @option options [String] :readonly (true) Whether to serialize this attribute
       #
       # @return [undefined]
       #
@@ -47,6 +50,8 @@ module Gecko
         model_name = association_name.to_s.singularize
         class_name = options[:class_name] || model_name.classify
         foreign_key = model_name.foreign_key.pluralize.to_sym
+        readonly    = options.key?(:readonly) ? options[:readonly] : true
+
         define_method association_name do
           ids = self.attributes[foreign_key]
           if ids.any?
@@ -55,7 +60,7 @@ module Gecko
             []
           end
         end
-        attribute foreign_key, Array[Integer]
+        attribute foreign_key, Array[Integer], readonly: readonly
       end
     end
   end