+++ test/record/fulfillment_test.rb
@@ -15,4 +15,13 @@ class Gecko::FulfillmentTest < Minitest::Test
   def test_initializes_record
     assert_instance_of(Gecko::Record::Fulfillment, @record)
   end
+
+  def test_serialization
+    json_keys = %w[
+      order_id shipping_address_id billing_address_id status exchange_rate
+      delivery_type tracking_number notes tracking_url tracking_company packed_at
+      shipped_at received_at receipt
+    ].map(&:to_sym)
+    assert_equal json_keys.sort, @record.as_json[:fulfillment].keys.sort
+  end
 end